+++ software/firmware/pru0-programmer/include/intelhex.h
@@ -13,9 +13,15 @@ typedef struct
 
 typedef enum
 {
-    IHEX_RET_OK   = 0u,
-    IHEX_RET_DONE = 1u,
-    IHEX_RET_ERR  = 2u
+    IHEX_RET_OK               = 0u,
+    IHEX_RET_DONE             = 1u,
+    IHEX_RET_ERR_REC_START    = 11u,
+    IHEX_RET_ERR_REC_CHECKSUM = 12u,
+    IHEX_RET_ERR_REC_END      = 13u,
+    IHEX_RET_ERR_LEN_EOF      = 21u,
+    IHEX_RET_ERR_LEN_ESAR     = 22u,
+    IHEX_RET_ERR_LEN_ELAR     = 24u,
+    IHEX_RET_ERR_TYPE_UNKNOWN = 31u,
 } ihex_ret_t;
 
 typedef struct
@@ -27,5 +33,6 @@ typedef struct
 
 int        ihex_reader_init(char *file_mem);
 ihex_ret_t ihex_reader_get(ihex_mem_block_t *block);
+uint32_t   ihex_get_line_number();
 
 #endif /* __PROG_INTELHEX_H_ */