+++ software/python-package/shepherd_sheep/cli.py
@@ -6,15 +6,18 @@ functionality to a command line user.
 
 """
 
+import atexit
 import signal
 import sys
 import time
+from datetime import datetime
 from pathlib import Path
 from types import FrameType
 from typing import TypedDict
 
 import click
 import gevent
+import yaml
 import zerorpc
 from shepherd_core import CalibrationCape
 from shepherd_core.data_models.task import ProgrammingTask
@@ -24,7 +27,6 @@ from shepherd_core.inventory import Inventory
 from typing_extensions import Unpack
 
 from . import __version__
-from . import run_programmer
 from . import run_task
 from . import sysfs_interface
 from .eeprom import EEPROM
@@ -35,6 +37,8 @@ from .shepherd_io import gpio_pin_nums
 from .sysfs_interface import check_sys_access
 from .sysfs_interface import disable_ntp
 from .sysfs_interface import reload_kernel_module
+from .usage_log import get_last_usage
+from .usage_log import usage_logger
 
 # allow importing shepherd on x86 - for testing
 try:
@@ -86,11 +90,17 @@ def cli(ctx: click.Context, *, verbose: bool, version: bool) -> None:
 
     if verbose:
         set_verbosity()
+
+    if ctx.invoked_subcommand and ctx.invoked_subcommand not in ["usage"]:
+        # this adds a usage-entry when sheep exits
+        atexit.register(usage_logger, datetime.now().astimezone(), ctx.invoked_subcommand)
+
     if version:
         log.info("Shepherd-Sheep v%s", __version__)
         log.debug("Python v%s", sys.version)
         log.debug("Click v%s", click.__version__)
-    check_sys_access()
+    if check_sys_access():
+        ctx.exit(1)
     if not ctx.invoked_subcommand:
         click.echo("Please specify a valid command")
 
@@ -152,13 +162,14 @@ def target_power(target_port: str, voltage: float, *, on: bool, gpio_pass: bool)
     type=click.Path(exists=True, readable=True, file_okay=True, dir_okay=False),
     default=Path("/etc/shepherd/config.yaml"),
 )
-def run(config: Path) -> None:
+@click.pass_context
+def run(ctx: click.Context, config: Path) -> None:
     reload_kernel_module()  # more reliable with fresh states
     disable_ntp()
     failed = run_task(config)
     if failed:
         log.debug("Tasks signaled an error (failed).")
-    sys.exit(int(failed))
+    ctx.exit(int(failed))
 
 
 @cli.group(
@@ -174,7 +185,9 @@ def eeprom() -> None:
     "cal-file",
     type=click.Path(exists=True, readable=True, file_okay=True, dir_okay=False),
 )
+@click.pass_context
 def write(
+    ctx: click.Context,
     cal_file: Path | None,
 ) -> None:
     cal_cape = CalibrationCape.from_file(cal_file)
@@ -184,7 +197,7 @@ def write(
             storage.write_calibration(cal_cape)
     except FileNotFoundError:
         log.error("Access to EEPROM failed (FS) -> is Shepherd-Cape missing?")
-        sys.exit(2)
+        ctx.exit(2)
 
 
 @eeprom.command(short_help="Read cape info and calibration data from EEPROM")
@@ -195,9 +208,20 @@ def write(
     default=None,
     help="If provided, calibration data is dumped to this file",
 )
-def read(cal_file: Path | None) -> None:
-    set_verbosity()
-
+@click.option(
+    "--revision",
+    "-r",
+    is_flag=True,
+    help="only output version of cape hardware on console",
+)
+@click.option(
+    "--full",
+    "-f",
+    is_flag=True,
+    help="output all fields on console",
+)
+@click.pass_context
+def read(ctx: click.Context, cal_file: Path | None, *, revision: bool, full: bool) -> None:
     try:
         with EEPROM() as storage:
             cal = storage.read_calibration()
@@ -205,20 +229,32 @@ def read(cal_file: Path | None) -> None:
         log.warning(
             "Reading from EEPROM failed (Val) -> no plausible data found",
         )
-        sys.exit(2)
+        ctx.exit(2)
     except FileNotFoundError:
         log.error("Access to EEPROM failed (FS) -> is Shepherd-Cape missing?")
-        sys.exit(3)
+        ctx.exit(3)
 
-    if cal_file is None:
-        log.info("Retrieved Cal-Data:\n\n%s", str(cal))
+    if revision:
+        log.info("%s", cal.cape.version)
+    elif cal_file is None:
+        _data = (
+            yaml.safe_dump(
+                cal.model_dump(exclude_unset=True, exclude_defaults=False),
+                default_flow_style=False,
+                sort_keys=False,
+            )
+            if full
+            else str(cal)
+        )
+        log.info("Retrieved Cal-Data:\n\n%s", str(_data))
     else:
         cal.to_file(cal_file)
 
 
 @cli.command(short_help="Start ZeroRPC Server")
 @click.option("--port", "-p", type=click.INT, default=4242)
-def rpc(port: int | None) -> None:
+@click.pass_context
+def rpc(ctx: click.Context, port: int | None) -> None:
     shepherd_io = ShepherdDebug()
     shepherd_io.__enter__()
     log.info("Shepherd Debug Interface: Initialized")
@@ -231,7 +267,7 @@ def rpc(port: int | None) -> None:
     def stop_server() -> None:
         server.stop()
         shepherd_io.__exit__()
-        sys.exit(0)
+        ctx.exit(0)
 
     gevent.signal_handler(signal.SIGTERM, stop_server)
     gevent.signal_handler(signal.SIGINT, stop_server)
@@ -291,7 +327,7 @@ def inventorize(output_path: Path) -> None:
     "--datarate",
     "-d",
     type=click.INT,
-    default=500_000,
+    default=200_000,
     help="Bit rate of Programmer (bit/s)",
 )
 @click.option(
@@ -306,15 +342,16 @@ def inventorize(output_path: Path) -> None:
     is_flag=True,
     help="dry-run the programmer - no data gets written",
 )
-def program(**kwargs: Unpack[TypedDict]) -> None:
+@click.pass_context
+def program(ctx: click.Context, **kwargs: Unpack[TypedDict]) -> None:
     protocol_dict = {
         "nrf52": ProgrammerProtocol.swd,
         "msp430": ProgrammerProtocol.sbw,
     }
     kwargs["protocol"] = protocol_dict[kwargs["mcu_type"]]
     cfg = ProgrammingTask(**kwargs)
-    failed = run_programmer(cfg)
-    sys.exit(int(failed))
+    failed = run_task(cfg)
+    ctx.exit(int(failed))
 
 
 @cli.command(
@@ -360,6 +397,16 @@ def blink(duration: int) -> None:
             time.sleep(0.125)
             dbg.select_port_for_io_interface(TargetPort.B)
             time.sleep(0.125)
+
+
+@cli.command(
+    short_help="Returns statistic about last usage: timestamp, total runtime, sub-command",
+    context_settings={"ignore_unknown_options": True},
+)
+@click.pass_context
+def usage(ctx: click.Context) -> None:
+    log.info(get_last_usage())
+    ctx.exit(0)
 
 
 if __name__ == "__main__":