+++ software/python-package/shepherd_watchdog/__init__.py
@@ -10,7 +10,7 @@ from types import TracebackType
 
 from typing_extensions import Self
 
-__version__ = "0.8.3"
+__version__ = "0.8.4"
 
 # Top-Level Package-logger
 log = logging.getLogger("ShpWatchdog")