+++ software/shepherd-calibration/shepherd_cal/__init__.py
@@ -2,15 +2,15 @@ from .calibration_plot import plot_calibration
 from .calibrator import Calibrator
 from .logger import activate_verbosity
 from .logger import logger
-from .profile import Profile
 from .profile_analyzer import analyze_directory
+from .profile_cape import ProfileCape
 from .profiler import Profiler
 
-__version__ = "0.8.3"
+__version__ = "0.8.4"
 
 __all__ = [
     "Calibrator",
-    "Profile",
+    "ProfileCape",
     "Profiler",
     "activate_verbosity",
     "analyze_directory",